Restructure repository and move to Automake.
Use Automake.
Use libtool for plugins.
Add ldns for rssm plugin.
Remove unused CWARN, OPTFLAGS.
Use Travis-CI.
