+++ .gitignore
@@ -1,7 +1,33 @@
 *.o
-config.status
+*.lo
+*.la
 config.log
+config.status
+stamp-h1
+ar-lib
+config.guess
+config.sub
+libtool
+ltmain.sh
+.deps
+.libs
+Makefile
+Makefile.in
+src/dnscap
+src/dnscap.1
+autom4te.cache
+Makefile.old
+aclocal.m4
+compile
+configure
+depcomp
+install-sh
+missing
+test-driver
 config.h
-dnscap
-dnscap.cat1
-*.swp
+config.h.in~
+m4/libtool.m4
+m4/ltoptions.m4
+m4/ltsugar.m4
+m4/ltversion.m4
+m4/lt~obsolete.m4
