+++ Makefile.am
@@ -1,0 +1,14 @@
+MAINTAINERCLEANFILES = $(srcdir)/Makefile.in \
+    $(srcdir)/src/config.h.in~ \
+    $(srcdir)/configure
+
+ACLOCAL_AMFLAGS = -I m4
+
+SUBDIRS = src plugins
+
+dist_doc_DATA = README.md LICENSE CONTRIBUTORS
+
+EXTRA_DIST = isc m4
+
+test:
+	@true
