+++ autogen.sh
@@ -1,0 +1,3 @@
+#!/bin/sh -e
+
+autoreconf --force --install --no-recursive
