+++ plugins/Makefile.am
@@ -1,0 +1,5 @@
+MAINTAINERCLEANFILES = $(srcdir)/Makefile.in
+
+SUBDIRS = pcapdump rssm txtout
+
+EXTRA_DIST = template
