+++ src/dnscap_common.h
